pyproject.toml
[project]
name = "constructio-mcp-server"
version = "0.1.0"
description = "Spec-first MCP server for Constructio"
requires-python = ">=3.11,<3.13"
dependencies = [
  "connexion[asgi,swagger-ui]>=3.0.0",
  "mcp>=1.0.0",
  "pydantic>=2.6.0",
  "google-cloud-storage>=2.16.0",
  "google-cloud-firestore>=2.16.0",
  "uvicorn>=0.27.0",
  "gcloud>=0.18.3",
  "fastmcp>=0.4.1",
  "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.12.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv.workspace]
members = ["constructio-mcp"]
